Update - Deprecate popups.js
