+++ source/js/statusControls.js
@@ -1,7 +1,7 @@
 import { deleteCharTracker, extensionSettings, setSaveStateFlag } from "../../index_old.js";
 import { chat, chat_metadata, extension_prompt_roles } from "../../../../../../script.js";
 import { t } from "../../../../../i18n.js";
-import { un_escapeNewlines } from "./popups.js";
+import { un_escapeNewlines } from "./popups_old.js";
 
 /** @type {Function} */
 toastr.error
